extract CSV source class
Split Source::CSV out into its own file
